# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RECALL API
  version: 0.1.0
  description: 个性化游戏生成与分享平台 MVP API。
servers:
  - url: http://127.0.0.1:8080
    description: Local development
paths:
...
  /api/v1/admin/invitation-codes:
    get:
      operationId: listRegistrationInvitations
      summary: 查询最近生成的邀请码及其使用状态
      description: 列表只返回脱敏邀请码，不返回完整邀请码或哈希。
      tags: [Admin Invitations]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 脱敏邀请码列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvitationListResponse"
        "401":
          $ref: "#/components/responses/APIError"
    post:
      operationId: createRegistrationInvitation
      summary: 生成一个只能成功注册一次的邀请码
      description: 完整邀请码仅在本次创建响应中返回一次，格式为 XXXX-XXXX。
      tags: [Admin Invitations]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "201":
          description: 新邀请码；这是唯一包含完整邀请码的响应
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedAdminInvitationResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
  /api/v1/admin/invitation-codes/{invitationId}:
    parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: string
          minLength: 26
          maxLength: 26
    delete:
      operationId: revokeRegistrationInvitation
      summary: 撤销一个尚未使用的邀请码
      tags: [Admin Invitations]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: 更新后的脱敏邀请码
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInvitationResponse"
        "401":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
...
  /api/v1/games/{gameId}/generation-runs:
    parameters:
      - $ref: "#/components/parameters/GameID"
    get:
      operationId: listGenerationRuns
      summary: 获取游戏的创建任务历史
      tags: [Generation Runs]
      responses:
        "200":
          description: 创建任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRunListResponse"
        "401":
          $ref: "#/components/responses/APIError"
    post:
      operationId: submitGenerationRun
      summary: 将当前游戏版本提交到 MySQL 任务队列
      tags: [Generation Runs]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitGenerationRequest"
      responses:
        "200":
          description: 相同幂等请求对应的已有任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRunResponse"
        "201":
          description: 新创建的排队任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRunResponse"
        "400":
          $ref: "#/components/responses/APIError"
        "403":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "422":
          $ref: "#/components/responses/APIError"
  /api/v1/games/{gameId}/generation-runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/GameID"
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: getGenerationRun
      summary: 获取创建任务当前阶段和进度
      tags: [Generation Runs]
      responses:
        "200":
          description: 创建任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRunResponse"
        "404":
          $ref: "#/components/responses/APIError"
  /api/v1/games/{gameId}/generation-runs/{runId}/cancel:
    parameters:
      - $ref: "#/components/parameters/GameID"
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: cancelGenerationRun
      summary: 取消排队任务或请求 Worker 安全停止运行中任务
      tags: [Generation Runs]
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/Origin"
      responses:
        "200":
          description: 更新后的任务状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationRunResponse"
        "404":
          $ref: "#/components/responses/APIError"
  /api/v1/admin/generation-runs:
    get:
      operationId: listAdminGenerationRuns
      summary: 获取不包含私人输入和素材的创建任务诊断列表
      tags: [Admin Generation Runs]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/GenerationRunStatus"
      responses:
        "200":
          description: 管理员创建任务列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGenerationRunListResponse"
        "401":
          $ref: "#/components/responses/APIError"
  /api/v1/admin/generation-runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunID"
    get:
      operationId: getAdminGenerationRun
      summary: 获取单条脱敏创建任务诊断
      tags: [Admin Generation Runs]
      responses:
        "200":
          description: 管理员创建任务详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGenerationRunResponse"
        "404":
          $ref: "#/components/responses/APIError"
...
components:
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
      example: http://127.0.0.1:5173
    GameID:
      name: gameId
      in: path
      required: true
      schema:
        type: string
...
    RunID:
      name: runId
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 由客户端为一次用户提交生成并在网络重试时复用，最长 128 字符。
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    APIError:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    AdminInvitation:
      type: object
      required: [id, codeHint, status, createdByAdmin, usedByCreatorId, usedByLoginId, usedAt, revokedAt, createdAt]
      properties:
        id:
          type: string
        codeHint:
          type: string
          pattern: "^••••-[2-9A-HJ-NP-Z]{4}$"
          description: 只保留末四位的脱敏标识。
        status:
          type: string
          enum: [unused, used, revoked]
        createdByAdmin:
          type: string
        usedByCreatorId:
          type: [string, "null"]
        usedByLoginId:
          type: [string, "null"]
        usedAt:
          type: [string, "null"]
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreatedAdminInvitation:
      allOf:
        - $ref: "#/components/schemas/AdminInvitation"
        - type: object
          required: [code]
          properties:
            code:
              type: string
              pattern: "^[2-9A-HJ-NP-Z]{4}-[2-9A-HJ-NP-Z]{4}$"
              description: 仅在创建成功时返回一次的完整邀请码。
    AdminInvitationResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/AdminInvitation"
        requestId:
          type: string
    CreatedAdminInvitationResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/CreatedAdminInvitation"
        requestId:
          type: string
    AdminInvitationListResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminInvitation"
        requestId:
          type: string
...
    SubmitGenerationRequest:
      type: object
      required: [versionId]
      properties:
        versionId:
          type: string
    GenerationRunStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    GenerationRunStage:
      type: string
      enum: [queued, transforming_images, saving_results, completed]
    GenerationRun:
      type: object
      required: [id, gameId, gameVersionId, attemptNumber, triggerType, status, stage, progress, errorCode, errorMessage, retryable, cancelRequested, createdAt, updatedAt, startedAt, completedAt]
      properties:
        id:
          type: string
        gameId:
          type: string
        gameVersionId:
          type: string
        attemptNumber:
          type: integer
          minimum: 1
        triggerType:
          type: string
          enum: [initial, user_retry]
        status:
          $ref: "#/components/schemas/GenerationRunStatus"
        stage:
          $ref: "#/components/schemas/GenerationRunStage"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 兼容字段；处理中为 0，最终状态为 100，不代表真实模型完成百分比。当前前端按 stage 映射为固定的展示百分比和四步文案。
        errorCode:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
        retryable:
          type: boolean
        cancelRequested:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
    AdminGenerationRun:
      allOf:
        - $ref: "#/components/schemas/GenerationRun"
        - type: object
          required: [executionCount, traceId, adminMessage, sanitizedDetails]
          properties:
            executionCount:
              type: integer
              minimum: 0
            traceId:
              type: string
            adminMessage:
              type: [string, "null"]
            sanitizedDetails:
              type: [object, "null"]
              additionalProperties: true
              description: 仅允许错误类型、诊断码、生成器版本等白名单字段。
    GenerationRunResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/GenerationRun"
        requestId:
          type: string
    GenerationRunListResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/GenerationRun"
        requestId:
          type: string
    AdminGenerationRunResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: "#/components/schemas/AdminGenerationRun"
        requestId:
          type: string
    AdminGenerationRunListResponse:
      type: object
      required: [data, requestId]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminGenerationRun"
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
        requestId:
          type: string
...